face-clustering-app/requirements.txt
gradio==4.44.0
opencv-python-headless==4.10.0.84
insightface==0.7.3
onnxruntime==1.19.2
hdbscan==0.8.38.post1
numpy==1.26.4
Pillow==10.4.0
scikit-learn==1.5.2
tqdm==4.66.5
